Moved some files around.
